Добавлен докер. Пофиксил ДБ и репы пользаков

+++ DeepLome.Models/Interfaces/Repositories/IUserRepository.cs
@@ -7,7 +7,7 @@ namespace DeepLome.Models.Interfaces.Repositories
         public void Add(User user);
         public void Delete(User user);
         public void Update(User user);
-        public User? GetById(int id);
+        public User? GetById(long id);
         public User? GetByName(string name);
         public IEnumerable<User> GetAll();
     }
